Move Settings struct to its own module
