+++ src/bundle/mod.rs
@@ -1,4 +1,4 @@
-use {PackageType, Settings};
+pub use self::settings::{PackageType, Settings};
 use std::path::PathBuf;
 
 mod common;
@@ -6,6 +6,7 @@ mod deb_bundle;
 mod ios_bundle;
 mod osx_bundle;
 mod rpm_bundle;
+mod settings;
 
 pub fn bundle_project(settings: Settings) -> ::Result<Vec<PathBuf>> {
     let mut paths = Vec::new();
